+++ metadata-ingestion/setup.py
@@ -763,8 +763,8 @@ entry_points = {
         "snowflake = datahub.ingestion.source.snowflake.snowflake_v2:SnowflakeV2Source",
         "snowflake-summary = datahub.ingestion.source.snowflake.snowflake_summary:SnowflakeSummarySource",
         "snowflake-queries = datahub.ingestion.source.snowflake.snowflake_queries:SnowflakeQueriesSource",
-        "superset = datahub.ingestion.source.superset:SupersetSource",
-        "preset = datahub.ingestion.source.preset:PresetSource",
+        "superset = datahub.ingestion.source.superset.superset:SupersetSource",
+        "preset = datahub.ingestion.source.superset.preset:PresetSource",
         "tableau = datahub.ingestion.source.tableau.tableau:TableauSource",
         "openapi = datahub.ingestion.source.openapi:OpenApiSource",
         "metabase = datahub.ingestion.source.metabase:MetabaseSource",
